CRUD de municipio

+++ app/Http/Controllers/ComunaController.php
@@ -111,13 +111,18 @@ class ComunaController extends Controller
      */
     public function destroy($id)
     {
-        $comuna =Comuna::find($id);
-        $comuna->delete();
-
+        $comuna = Comuna::find($id);
+    
+        if ($comuna) {
+            $comuna->delete();
+        }
+    
         $comunas = DB::table('tb_comuna')
-        ->join('tb_municipio','tb_comuna.muni_codi','=','tb_municipio.muni_codi')
-        ->select('tb_comuna.*',"tb_municipio.muni_nomb")
-        ->get();
-        return view('comuna.index',['comunas'=>$comunas]);
+            ->join('tb_municipio', 'tb_comuna.muni_codi', '=', 'tb_municipio.muni_codi')
+            ->select('tb_comuna.*', 'tb_municipio.muni_nomb')
+            ->get();
+    
+        return view('comuna.index', ['comunas' => $comunas]);
     }
+    
 }
